added performance view and settings view toggle
