+++ src/App.css
@@ -448,6 +448,20 @@ li {
     grid-template-columns: minmax(0, 0.7fr) minmax(0, 1fr) minmax(0, 1fr);
 }
 
+.dashboard-grid-settings{
+    width: 96%;
+    height: 94%;
+    display: grid;
+    grid-template-areas: 'header header header'
+    'menu setting setting'
+    'menu setting setting';
+    gap: 2%;
+    padding-left: 2%;
+    padding-top: 2%;
+    grid-template-rows: minmax(0, 0.4fr) minmax(0, 1fr) minmax(0, 1fr);
+    grid-template-columns: minmax(0, 0.7fr) minmax(0, 1fr) minmax(0, 1fr);
+}
+
 .dashboard-header{
     background-color: #FFFFFF;
     grid-area: header;
@@ -664,6 +678,14 @@ li {
     display: flex;
     width: 90%;
     overflow-x: auto;
+}
+
+.dashboard-setting{
+    background-color: #FFFFFF;
+    grid-area: setting;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
 }
